# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Platform for online internships",
    "version": "v1"
  },
  "paths": {
...
    "/api/competencies/": {
      "get": {
        "summary": "Get Competencies",
        "operationId": "get_competencies_api_competencies__get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_Competence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_competencies__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/competencies/{competence_id}": {
      "get": {
        "summary": "Get Competence",
        "operationId": "get_competence_api_competencies__competence_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_competencies__competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Patch Competence",
        "operationId": "patch_competence_api_competencies__competence_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/competencies/": {
      "get": {
        "summary": "Get Course Competencies",
        "operationId": "get_course_competencies_api_courses__course_id__competencies__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_CourseCompetence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_courses__course_id__competencies__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCompetence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/courses/{course_id}/competencies/{course_competence_id}": {
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_courses__course_id__competencies__course_competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Competence Id",
              "type": "integer"
            },
            "name": "course_competence_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/subdivisions/{subdivision_id}/posts/{post_id}/competencies/": {
      "get": {
        "summary": "Get Post Competencies",
        "operationId": "get_post_competencies_api_subdivisions__subdivision_id__posts__post_id__competencies__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_PostCompetence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_subdivisions__subdivision_id__posts__post_id__competencies__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCompetence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/subdivisions/{subdivision_id}/posts/{post_id}/competencies/{post_competence_id}": {
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_subdivisions__subdivision_id__posts__post_id__competencies__post_competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Competence Id",
              "type": "integer"
            },
            "name": "post_competence_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Competence": {
        "title": "Competence",
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "CourseCompetence": {
        "title": "CourseCompetence",
        "required": [
          "id",
          "course_id",
          "course_name",
          "competence_id",
          "competence_name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "course_name": {
            "title": "Course Name",
            "type": "string"
          },
          "competence_id": {
            "title": "Competence Id",
            "type": "integer"
          },
          "competence_name": {
            "title": "Competence Name",
            "type": "string"
          }
        }
      },
...
      "CreateCompetence": {
        "title": "CreateCompetence",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
      "CreateCourseCompetence": {
        "title": "CreateCourseCompetence",
        "required": [
          "competence_id"
        ],
        "type": "object",
        "properties": {
          "competence_id": {
            "title": "Competence Id",
            "type": "integer"
          }
        }
      },
      "CreatePostCompetence": {
        "title": "CreatePostCompetence",
        "required": [
          "competence_id"
        ],
        "type": "object",
        "properties": {
          "competence_id": {
            "title": "Competence Id",
            "type": "integer"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "LimitOffsetPage_Competence_": {
        "title": "LimitOffsetPage[Competence]",
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competence"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1.0,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0.0,
            "type": "integer"
          }
        }
      },
      "LimitOffsetPage_CourseCompetence_": {
        "title": "LimitOffsetPage[CourseCompetence]",
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseCompetence"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1.0,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0.0,
            "type": "integer"
          }
        }
      },
...
      "LimitOffsetPage_PostCompetence_": {
        "title": "LimitOffsetPage[PostCompetence]",
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostCompetence"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1.0,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0.0,
            "type": "integer"
          }
        }
      },
...
      "PatchCompetence": {
        "title": "PatchCompetence",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "PostCompetence": {
        "title": "PostCompetence",
        "required": [
          "id",
          "post_id",
          "post_name",
          "competence_id",
          "competence_name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "post_name": {
            "title": "Post Name",
            "type": "string"
          },
          "competence_id": {
            "title": "Competence Id",
            "type": "integer"
          },
          "competence_name": {
            "title": "Competence Name",
            "type": "string"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}